@@ -1,8 +1,8 @@
 package com.simple.logging.configuration;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import javax.validation.constraints.NotNull;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import jakarta.validation.constraints.NotNull;
 import org.springframework.web.method.HandlerMethod;
 import org.springframework.web.servlet.DispatcherServlet;
 import org.springframework.web.servlet.HandlerExecutionChain;
@@ -38,23 +38,26 @@ public class LoggableDispatcherServlet extends DispatcherServlet {
     private final String logFilePath;
     private final String charset;
     private final Integer maxCacheHistoryLogs;
+    private final String applicationName;
 
     /**
      * Constructs a new LoggableDispatcherServlet with specified logging configurations.
      *
-     * @param maxFileSize         the maximum size of the log file in megabytes.
-     * @param maxStringSize       the maximum size of the request/response body to be logged in megabytes.
-     * @param logFilePath         the directory path where log files will be stored.
-     * @param charset             the character encoding to be used for logging.
-     * @param maxCacheHistoryLogs the maximum number of logs to be cached in memory.
+     * @param maxFileSizeMb            the maximum size of the log file in megabytes.
+     * @param maxStringSizeMb          the maximum size of the request/response body to be logged in megabytes.
+     * @param logFilePath              the directory path where log files will be stored.
+     * @param charset                  the character encoding to be used for logging.
+     * @param maxCacheHistoryLogs      the maximum number of logs to be cached in memory.
+     * @param applicationName          name of your application.
      */
-    public LoggableDispatcherServlet(int maxFileSize, int maxStringSize, String logFilePath,
-                                     String charset, Integer maxCacheHistoryLogs) {
-        this.maxFileSizeMb = maxFileSize * 1024 * 1024; // Convert MB to bytes
-        this.maxStringSizeMb = maxStringSize * 1024 * 1024;
+    public LoggableDispatcherServlet(Integer maxFileSizeMb, Integer maxStringSizeMb, String logFilePath,
+                                     String charset, Integer maxCacheHistoryLogs, String applicationName) {
+        this.maxFileSizeMb = maxFileSizeMb * 1024 * 1024; // Convert MB to bytes
+        this.maxStringSizeMb = maxStringSizeMb * 1024 * 1024;
         this.logFilePath = logFilePath;
         this.charset = charset;
         this.maxCacheHistoryLogs = maxCacheHistoryLogs;
+        this.applicationName = applicationName;
         setupLogger();
     }
 
@@ -72,7 +75,7 @@ public class LoggableDispatcherServlet extends DispatcherServlet {
             // Define log filename with date
             DateTimeFormatter dtf = DateTimeFormatter.ofPattern("yyyy-MM-dd");
             String dateTime = LocalDate.now().format(dtf);
-            Path logFile = logsPath.resolve("application-" + dateTime + ".log");
+            Path logFile = logsPath.resolve(applicationName + "-" + dateTime + ".log");
 
             // Create FileHandler with size limit and rotating file pattern
             FileHandler fileHandler = new FileHandler(logFile.toString(), maxFileSizeMb, 1, true);
@@ -141,7 +144,7 @@ public class LoggableDispatcherServlet extends DispatcherServlet {
 
         // Delete first element if the list is overflown
         if (PayloadHistory.viewLogs().size() >= maxCacheHistoryLogs) {
-            PayloadHistory.viewLogs().removeFirst();
+            PayloadHistory.viewLogs().remove(0);
         }
 
         PayloadHistory.addLog(log);
@@ -189,7 +192,7 @@ public class LoggableDispatcherServlet extends DispatcherServlet {
             String jsonStringFromByteArray = new String(byteArray, StandardCharsets.UTF_8);
             String payloadMarker = isPayloadResponse ? "RESPONSE BODY: {0}" : "REQUEST BODY: {0}";
             if (!jsonStringFromByteArray.isBlank()) {
-              LOGGER.log(Level.INFO, payloadMarker, jsonStringFromByteArray);
+                LOGGER.log(Level.INFO, payloadMarker, jsonStringFromByteArray);
             }
             // Check whether the payload is a response or a request
             if (isPayloadResponse) {
